% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\alias{download_peptides}
\alias{download_nucleotides}
\title{Download a dataset}
\usage{
download_peptides(x)

download_nucleotides(x)
}
\arguments{
\item{x}{OneKP object}
}
\value{
character vector of paths to the files that were downloaded
}
\description{
These functions will return all files in the OneKP object of the given type
(e.g. protein FASTA files for get_peptides). If you do not want to retrieve
all these files (there are over a thousand), then you should filter the
OneKP object first, using the \code{filter_by_*} functions.
}
