% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter}
\alias{filter}
\alias{filter_by_code}
\alias{filter_by_node}
\alias{filter_by_species}
\title{Filter a OneKP object}
\usage{
filter_by_code(x, code)

filter_by_node(x, node)

filter_by_species(x, species)
}
\arguments{
\item{x}{OneKP object}
}
\value{
OneKP object
}
\description{
Filter a OneKP object
}
