% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/onekp.R
\docType{data}
\name{onekp}
\alias{onekp}
\alias{onekp-package}
\title{OneKP metadata file}
\format{OneKP object}
\usage{
onekp
}
\description{
The object stored here should be exactly the same as the object returned
from \code{retrieve_onekp()}. It is stored here for convenience and to save
time in examples (\code{retrieve_onekp} takes around 30 seconds to run).

The 1000 Plants Initiative (www.onekp.com) has sequenced the transcriptomes
of over 1000 plant species. This package allows these sequences and metadata
to be retrieved and filtered by code, species or recursively by clade.
Scientific names and NCBI taxonomy IDs are both supported.
}
\section{Main Functions}{


\code{retrieve_onekp} - retrieve all 1KP metadata

\code{filter_by_code} - filter metadata by 1KP code

\code{filter_by_clade} - filter metadata by clade

\code{filter_by_species} - filter metadata by species

\code{download_peptides} - get protein sequences linked to metadata

\code{download_nucleotides} - get DNA sequences linked to metadata
}

\section{Author(s)}{


Zebulun Arendsee <email: zbwrnz@gmail.com>
}

\section{Bug Reports}{


Any bugs or issues can be reported at <https://github.com/ropensci/onekp/issues>
}

\keyword{datasets}
