% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.OneKP}
\alias{print.OneKP}
\title{OneKP print generic function}
\usage{
\method{print}{OneKP}(x, ...)
}
\arguments{
\item{x}{OneKP object}

\item{...}{Additional arguments (unused)}
}
\description{
OneKP print generic function
}
