% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\alias{download_peptides}
\alias{download_nucleotides}
\title{Download a dataset}
\usage{
download_peptides(x, dir = file.path(tempdir(), "peptides"), absolute = FALSE)

download_nucleotides(
  x,
  dir = file.path(tempdir(), "nucleotides"),
  absolute = FALSE
)
}
\arguments{
\item{x}{OneKP object}

\item{dir}{Directory in which to store the downloaded data}

\item{absolute}{If TRUE, return absolute paths (default=FALSE)}
}
\value{
character vector of paths to the files that were downloaded
}
\description{
These functions will return all files in the OneKP object of the given type
(protein or DNA FASTA files for \code{download_peptides} and
\code{download_nucleotides}, respectively). If you do not want to retrieve
all these files (there are over a thousand), then you should filter the
OneKP object first, using the \code{filter_by_*} functions.
}
\examples{
\dontrun{
data(onekp)

# Filter by 1KP code (from `onekp@table$code` column)
seqs <- filter_by_code(onekp, c('URDJ', 'ROAP'))

# Download FASTA files to temporary directory 
download_peptides(seqs)
download_nucleotides(seqs)
}
}
