% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter}
\alias{filter}
\alias{filter_by_code}
\alias{filter_by_ancestor_name}
\alias{filter_by_ancestor_id}
\alias{filter_by_species}
\alias{filter_by_taxid}
\title{Filter a OneKP object}
\usage{
filter_by_code(x, code)

filter_by_ancestor_name(x, names)

filter_by_ancestor_id(x, ids)

filter_by_species(x, names)

filter_by_taxid(x, ids)
}
\arguments{
\item{x}{OneKP object}

\item{code}{character vector of 1KP IDs (e.g. URDJ)}

\item{names}{vector of scientific names (}

\item{ids}{vector of NCBI taxonomy IDs}
}
\value{
OneKP object
}
\description{
Filter a OneKP object
}
\examples{
onekp <- retrieve_oneKP()
onekp

# filter by 1KP ID
filter_by_code(onekp, c('URDJ', 'ROAP'))

# filter by species name
filter_by_species(onekp, 'Pinus radiata')

# filter by species NCBI taxon ID
filter_by_taxid(onekp, 3347)

# filter by ancestor name scientific name
filter_by_ancestor_name(onekp, 'Brassicaceae')

# filter by ancestor NCBI taxon ID
filter_by_ancestor_id(onekp, 3700)
}
