% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter}
\alias{filter}
\alias{filter_by_code}
\alias{filter_by_clade}
\alias{filter_by_species}
\title{Filter a OneKP object}
\usage{
filter_by_code(x, code)

filter_by_clade(x, clade)

filter_by_species(x, species)
}
\arguments{
\item{x}{OneKP object}

\item{code}{character vector of 1KP IDs (e.g. URDJ)}

\item{clade}{vector of clade-level NCBI taxonomy IDs or scientific names}

\item{species}{vector of species-level scientific names or NCBI taxonomy IDs}
}
\value{
OneKP object
}
\description{
Filter a OneKP object
}
\examples{
onekp <- retrieve_oneKP()
onekp

# filter by 1KP ID
filter_by_code(onekp, c('URDJ', 'ROAP'))

# filter by species name
filter_by_species(onekp, 'Pinus radiata')

# filter by species NCBI taxon ID
filter_by_species(onekp, 3347)

# filter by clade name scientific name
filter_by_clade(onekp, 'Brassicaceae')

# filter by clade NCBI taxon ID
filter_by_clade(onekp, 3700)
}
